+++ service/matching/matchingEngine.go
@@ -26,10 +26,6 @@ import (
 	"math"
 	"sync"
 
-	"github.com/uber/cadence/common/service/dynamic"
-
-	"github.com/pborman/uuid"
-	"github.com/uber-common/bark"
 	h "github.com/uber/cadence/.gen/go/history"
 	m "github.com/uber/cadence/.gen/go/matching"
 	workflow "github.com/uber/cadence/.gen/go/shared"
@@ -39,6 +35,10 @@ import (
 	"github.com/uber/cadence/common/logging"
 	"github.com/uber/cadence/common/metrics"
 	"github.com/uber/cadence/common/persistence"
+	"github.com/uber/cadence/common/service/dynamicconfig"
+
+	"github.com/pborman/uuid"
+	"github.com/uber-common/bark"
 )
 
 // Implements matching.Engine
